sugar-session: disable Metacity mouse button modifiers, OLPC #11781
Metacity provides "mouse button modifier" functionality which lets
you drag activity windows around the screen, amongst other things.
This is something we want to disable, similar to how we disable
metacity's key bindings.

In recent metacity releases, Alt is the default mouse button modifier.
We especially want to disable this functionality in this case because
Sugar uses Alt for other things, such as launching a new activity
instance (rather than a journal entry) from the home screen.

This patch depends on the metacity commit 86fd12cbf06f693405650b9d6f7c6234757cc354
which adds a new 'metacity-message' option to "enable-mouse-button-modifiers"
and to "disable-mouse-button-modifiers", in both cases it has to be set to
true. For example, this patch has been added to metacity-2.34.3-2.fc17 in
Fedora. Running Sugar with a metacity version that does not have this patch
does work fine, we exit gracefully and display a warning that the
"disable-mouse-button-modifiers" message could not be applied.

There are two ways to verify this:
- open the Terminal Activity, hold the alt-key down and try to
  click-drag the window ---> dragging the window around should not
  be possible
- hold the alt key down and start an Activity from the home
  screen ---> a new instance should be created rather than resuming
  an old one

Signed-off-by: Daniel Drake <dsd@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>