Enable gnome-keyring-daemon to start inside emulated session
We have the following situation: there is an instance of gnome-keyring-daemon
running on the machine. But sugar-emulator can't access this instance of
gnome-keyring-daemon, as it is outside the scope of sugar-emulator and not
registered with the D-Bus session bus started by sugar-emulator.
D-Bus tries to start a new instance of gnome-keyring-daemon, which fails due
to a set of environment variables related to gnome-keyring-daemon.  These
essentially make gnome-keyring-daemon try to initialize the existing instance
of gnome-keyring-daemon outside the scope of sugar-emulator.
By removing these environment variables, D-Bus can start a new instance of
gnome-keyring-daemon and register with the D-Bus session.

See
https://bugzilla.gnome.org/show_bug.cgi?id=628302
and
https://live.gnome.org/GnomeKeyring/RunningDaemon

for more information.

[changed description and comment]
Signed-off-by: Caspar Bothmer <caspar@activitycentral.com>
Reviewed-by: Caspar Bothmer <caspar@activitycentral.com>
Acked-by: Sascha Silbe <silbe@activitycentral.com>