Frame: reveal Palette on left click for device, friends and clipboard items
In previous design discussions we agreed on making the Palette appear
immediately on left click. Actually some device icons like Speech or
Speaker have been already doing this.

The device icons that change to that behavior with this patch
are: all the network items, the volume item and the battery item.
All the items are listening to the 'button-release-event'
and do not propagate the event further after revealing the Palette.

The Friends items behave the same as the device icons technically
(both are TrayIcons) and UI-wise.

We could think about handling the left click directly in the TrayIcon
there are a few exceptions like the touchpad device icon and the volume
icons which do have a primary action on left click.

As well the clipboard icons will reveal now the Palette on left
click. This is also true when selecting a clipboard item. Technically we
have to listen to the 'button-release-event' of the child widget
as the RadioToolButton does only emit a clicked signal. Using this
would colide with the state changes of the clipboard items as those
emit a signal [1]. We do let the event propagate further
(return False) so that the state change when selecting another
item is picked up.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>

[1] http://developer.gnome.org/gtk3/3.4/GtkToggleToolButton.html#gtk-toggle-tool-button-set-active