Journal, no-hippo: set white background for no-matching-entry message, SL #3808
The 'no-matching-entry' message in the Journal listview is
grey after the hippo removal.

Add a gtk.EventBox because the gtk.VBox doesn't have a
gtk.gdk.Window associated with it and the background
can't be set otherwise. See d9fbf9d for the same resolution
of the issue and some background info [1].

[1] https://bugzilla.gnome.org/show_bug.cgi?id=678790

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>