Use the gdk window to get the depth of the preview pixmap - SL#3804
This prevents harcoding the depth, and uses the one from the gdk
window.  For this, the preview has to be done after the window is
realized, so the code is wrapped in the expose-event callback.

Also add a white background to the preview, otherwise artifacts appear
for previews smaller than the preview box.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>