SpreadLayout: do not remove items from the grid if there is none yet, SL #3814
A grid is added to the ViewContainer when the children
are allocated, we can not do it earlier because we need to
pass the allocation details. The allocation happens when
the children needs to be shown. Therefore if the container
and it's children have not been shown yet, removing the
child from the grid would fail.

This patch does check if a grid exists already before trying
to remove items from it.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>