Clipboard, drag&drop: keep notification icon up-to-date, SL #3818
When we drag&drop an object we do set the completion percentage to 100
when the drop begins [1]. This creates the Notification icon right at the
beginning when the data type has not been determined yet [2]. The
Clipboard icon itself we keep on updating when the state changes [3].

This patch does update the Notification icon on state changes like we do
for the Clipboard icon. This patch needs Manuel's fix for SL #3811
in order to be verified.

[1] http://git.sugarlabs.org/sugar/mainline/blobs/master/src/jarabe/frame/clipboardtray.py#line184
[2] http://git.sugarlabs.org/sugar/mainline/blobs/master/src/jarabe/frame/clipboardicon.py#line137
[3] http://git.sugarlabs.org/sugar/mainline/blobs/master/src/jarabe/frame/clipboardicon.py#line123

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>