Frame: Drop support for XdndDirectSave in the clipboard
It was added to allow dragging images from the old mozilla based
web browser to the frame. The mozilla side of the patch was never
upstreamed and anyway the new webkit based browser doesn't support
the protocol.

This is preliminary to the gtk3 port, we are trying to get rid of
window.property_change calls which are not working with introspection.

Signed-off-by: Daniel Narvaez <dwnarvaez@gmail.com>
Acked-by: Manuel Quiñones <manuq@laptop.org>