Run sugar-convert.sh for automatic conversion from sugar.* to sugar3.* imports
Run sugar-convert.sh [1] for automatic conversion from sugar.* imports to
sugar3.* imports. This is only on a best-effort basis; the code will
be in a broken state after this patch and need to be fixed manually.

The purpose of committing the intermediate, non-working output is to make it
reproducible. It's impractical to manually review the changes.

[1] http://dev.laptop.org/~manuq/sugar-convert.sh

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>