All: Chain up to constructor of derived class
The pygi-convert.sh script does replace the chaining to the
constructor of the derived class and chains up to the GObject
constructor, those introduced missing imports and a few more
issues with arguments passing.

In general it is clearer to chain up to the constructor of
the derived class so while fixing the issues introduced by the
script we did switch to that behavior in all of the shell.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>