Add feedback in Home View, List View when there are no matching entries - SL #3838
The scrolled window widget that contains the list of activities is
replaced by a message box when there are no matching entries.  This is
the same as the Journal does.  An alert can be packed also, so we use
pack_end to add the message box or the list of activities again.

Use Gtk.TreeModel.iter_n_children() [1] to count the matches, note
that this work only for tree model lists.

The message box widget is copied from the Journal listview.  Is done
in a separate class and can be generalized later.  The only important
change here is the icon colors: to make the 'system-search' icon look
ok, the stroke color and fill color were switched.

[1] http://developer.gnome.org/gtk3/stable/GtkTreeModel.html#gtk-tree-model-iter-n-children

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>