Port Volume palette that has mixed content to new API
This fixes VolumePalette that is used in the Journal, and in the
volume device icon.

Had to set the content box (the one passed to set_content) a property
of the palette.  Otherwise the "Show Contents" can't be added from the
device icon.  Previously the palette.menu attribute was used for that.

Also I had to reorder the "Show Contents" item because pack_end nor
pack_start didn't pack it to the correct place.

Note that the separator packed after the Remove item is not visible.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>