Fix external media mounting, SL #3911
The GVolumeMonitor used for mounting devices must not drop out
of scope, otherwise it will be destroyed, and no signals will be
receieved. Move it into global scope.

Update mount and unmount calls for introspection.

Remove dead code from volumestoolbar.

The user data argument that needs to be passed to the mount/unmount
method is discussed at [1].

[1] https://mail.gnome.org/archives/python-hackers-list/2012-September/msg00009.html

Signed-off-by: Daniel Drake <dsd@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>