Finish port of palettes in network device icon
Started in aaa963cf with the port of wireless.  The pending one was
GsmPalette.  WiredPalette doesn't need to be changed, it doesn't pack
MenuItems, just sets the content with set_content, is all
informational.  And device type OLPC mesh uses the same palette as
wireless, already ported.

- use PaletteMenuItem instead of MenuItem
- fixed the calls to set_label of the menu item
- use update_state instead of unexistent set_state

This needs to be properly tested with a modem connection.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>