Favorites view: get back hover feedback in activity icons - SL #3894
Since de-hippo, our icons are implemented as Gtk.EventBox.  If a frame
is drawn in the do_draw method using Gtk.render_frame, it can be
styled in the theme, setting border rules.

Listening to the enter-notify and leave-notify events, we can set the
prelight state to the eventbox.  Then in the theme we can style that
state.

This is convenient to get back the grey rectangle in the favorites
view when an activity icon is hovered.  The previous cairo drawing
that this patch removes is broken, and seems more complicated than
using Gtk.render_frame.  Also is better to do the styling in the
theme.

This patch will be accompanied with an artwork patch that adds the
styling.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>