Request AP properties with dbus ByteArrays (#4031)
As we deal with the ssid property as a string, we need to explicitly
request byte array data as a ByteArray as is done in other places.

This wasn't a problem before the upgrade to dbus-1.6.8; its likely
that the byte_arrays setting was questionably inherited from other
places where we subscribe to this signal. But now it seems like
each call site is independent, which makes sense, and each one must
request byte_arrays if that is the format it wants to work with.

Signed-off-by: Daniel Drake <dsd@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>