NetworkView: add the custom SugarPalette instead of the GTK.Menu based one, part of SL #3992
The Gtk.Menu triggers a Palette popdown in a long press (and actually as well
a long click). There is not realy a way around that so we go with our custom
Palettes.

This adds the custom Palette for the SugarAdhocView, the WirelessNetworkView
and the OlpcMeshView.

Use the toggle_palette property from the invoker in order to popup/popdown
Palette on left click or tap.

Fix appearance of AdHoc icons as well, error introduced
in 02218ea3e0c4ac2edaa8b08dbc066f2d22f29e8a.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>