Frame: use the new toggle_palette invoker property, SL #4065
For the device icons, the friends icons and the clipboard ones
we use the toggle_palette invoker property to achieve the
following behavior: the Palette will popup/popdown on button
left click/touch tap.

The property has been added in c4165967d5e05607db8b3e0969b516da87855431
to the toolkit.

For the network device icons OlpcMeshDeviceView and WirelessDeviceView
we do have to set an extra property to make this work. They are
ToolButtons and not TrayIcons like the other icons. By default the
toolbutton does popdown the Palette on a click.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>