BuddyMenu: switch from GTK.Menu to custom menu, part of SL #3992
The Gtk.Menu triggers a Palette popdown in a long press (and actually as well
a long click). There is not realy a way around that so we go with our custom
Palettes.

Use the toggle_palette property from the invoker in order to popup/popdown
Palette on left click or tap.

Fixme: we need to make the Palette widget accessible through the Palette
API.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>