Move Wireless Key dialog to an upper screen position, SL #4050
In ebook mode the dialog would be obscured by the OSK
otherwise.

The y position is style.GRID_CELL_SIZE below the toolbar. The x
value is centered based on the screen width. Using gtk-window-get-size
we know the size of the dialog before it is on screen to do the
calculation [1].

[1] http://developer.gnome.org/gtk3/3.4/GtkWindow.html#gtk-window-get-size

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>