Owner Icon: deactivate mouse over for the fist time - SL #4221
The icons in all shell views have a grey rectangle background as mouse
over feedback, but as the mouse pointer is set in the middle of the
view when Sugar is initializated, we get the owner icon hovered, which
we don't want.  This commit overrides that behaviour, in a similar way
that is being done for the palette, in order to don't trigger it the
fist time.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>