Adapt to PaletteMenu changes for popdown on item activation, SL #4191 and SL #4253
The toolkit now popdown the Palette when an
item is activated. We do not have to do it manually
any more.

The Control Panel item code was setting the panel
as transient on itself, this fails as the Palette
vanishes after the item is activated. We are doing
fine without setting the Control Panel transient on
the Palette, which is wrong in any case.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>