Journal: fix copy to clipboard functionality, part of #4307
gtk_clipboard_set_with_data expects us to pass a data argument [1]. As the
tempory file path we do copy to the clipboard is already part of the model
we can set this to None as we can access that in the get_func and
clear_func.

The targets have to be of type GtkTargetEntry.

[1] http://developer.gnome.org/gtk3/3.4/gtk3-Clipboards.html#gtk-clipboard-set-with-data
[2] http://developer.gnome.org/gtk3/3.4/gtk3-Selections.html#GtkTargetEntry

Signed-off-by: Simon Schampijer <simon@laptop.org>
Reviewed-by: Daniel Drake <dsd@laptop.org>