Move metacity invocation into the python code
So that the sugar script only setup environment variables
and all the rest is in main.py. This avoids the confusion about
putting stuff in the one or the other script. Ideally we would
get rid of the bash script completely, but it would not be
possible to source the debug and i18n files from python.

Acked-by: Simon Schampijer <simon@laptop.org>