Rework how we select the Xcursor theme
Metacity reads the default theme from gsettings and we were trying
to override it at lower level, causing race conditions. Now we are
just setting the cursor-theme gsetting, so that metacity setup the
Sugar theme for us.

Distributions which are shipping both GNOME and Sugar should make
sure to set XDG_CONFIG_HOME for either of the desktops to point
to a non default location, so that the settings are not conflicting.
It's necessary in this case if you don't want to get the Sugar
cursors in GNOME, but it seems like a sane approach in general. We
are sharing several components with GNOME which are (or might in
the future) use GNOME gsettings.

Acked-by: Simon Schampijer <simon@laptop.org>