ObjectChooser: make ObjectChooser transient for the parent window
There are two set-transient-for for methods: one that expects the
window and parent parameter to be of type GtkWindow [1] and one that
expects them to be of type GdkWindow [2].

The parent window we get for the object chooser is of type GdkWindow
as we use GdkX11.X11Window.foreign_new_for_display to construct it
from the xid passed from the toolkit.

This patch changes the code to use gdk_window_set_transient_for in
this case.

[1] http://developer.gnome.org/gtk3/3.4/GtkWindow.html#gtk-window-set-transient-for
[2] http://developer.gnome.org/gdk3/stable/gdk3-Windows.html#gdk-window-set-transient-for

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>