Journal toolbar: add tooltip to favorite filter
This patch adds the tooltip 'Favorite entries' to the favorite filter
button in the Journal toolbar. The button is a ToggleToolButton and
with the recent change in the toolkit-gtk3 63b8e87b1a99a854e9adbb1579b1e05244d2dc4
we do hide the tooltip when the button is clicked or touched.

This adds a new string, therefore this patch is only applied to
master. The addition and string has been discussed with Gary
and Manuel.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>