Control Panel: handle switching between landscape and portrait mode, SL #4444
When the control panel is opened for the first time, a table is
generated with a number of columns based on screen width. If the
screen is subsequently rotated, the column number is no longer
correct. In addition, the scrolled window holding the table needs to
be resized so that it fits on the screen.

This patch removes the section icons and sets it up again to
recompute the number of columns and to resize the window and
it's containing table.

Signed-off-by: Walter Bender <walter.bender@gmail.com>
Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>