documents dir: handle cases where get_documents_path returns None, SL #4380
I came across this when the Palettes in the Journal where not able
to be displayed because the code to add the copy to documents option
was failing hard because get_documents_path returned None.

The root cause on my system is that I have wiped the Documents folder and
xdg-user-dir returns $HOME as a consequence which we ommit in
get_documents_path.

This patch does gracefully handle those cases in a few places, including
the Palettes. In the volumes toolbar we were already doing so.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>