Don't lint config.py
It's autogenerated and the paths can be an arbitrary. So it can
overflow line length depending on where you install it. Fortunately
there is not much to lint in config.py.

I'm doing this using flake8 because the pep8 exclude option is not
very good and doing this in the code is nicer anyway. flake8 runs
both pyflakes and pep8, adding extension support and the ability
to exclude lines and entire files.