Bundle registry: Cleanup/clarify bundle installation
Management of bundle installs/upgrades was a mess, this attempts to
clean it up and sort it out.

It was not clear under which circumstances a bundle should be installed
via the bundle object vs via the bundle registry. The logical answer here
is that it should be done by the registry, so document why that is the case.

In bundleregistry, upgrade() seemed like it should be used for bundle
updates, but actually install() duplicated some of that logic in a slightly
different way. Touch up the upgrade logic in install(), remove upgrade(),
and add some documentation to make the situation clear.

The install() API differed between JournalEntryBundle, ActivityBundle and
ContentBundle. Eliminate the differences. JEB uid's are now stored at
initialization time, and the Bundle classes themselves choose the install
path (this does not vary when activities are installed this way).

The journal had two codepaths that dealt with activity installation/upgrade
(on arrival of the bundle in the datastore, and on resume)); unite them into
one and update for the above changes.