bundleregistry: Fix conflict between file monitoring and its own operations (#3707)
When an activity is installed via bundleregistry, it is effectively
registered twice: once by bundleregistry after installing it, and again
when the file operations monitor picked up the event.

This lead to error messages in the logs and occasional duplication
of activity icons in the home view.

Refactor the code slightly so that it is resilient and accepting that
some activities may try to be registered twice. To avoid the duplication
of icons, be more careful to only emit the bundle-added signal when we
really add the bundle.

Also clean up the installation of MIME types. The ActivityBundle class
does this, so there is no need for bundleregistry to do so.