Store content bundles in bundle registry
Previously, only activity bundles were maintained in the bundle registry.
Extend this to cover all kinds of "trackable" bundles, which in reality
just means the addition of content bundles.

This is needed in order to clean up the code and bundle API (as has happened
here) and to allow content bundle updates through the updater control
panel.