Add online service support
This patch adds a new section to sugar/src/jarabe to manage web services.

The relevant feature request is [1].

Adds a new web section of jarabe is for a prototype of a new online
account management class that manages the services described above.
The services themselves are intended to be kept in subdirectors of
sugar/extensions/web and are the subject of a separate patch. Also
in a separate patch will be a UI extention to the Journal toolbar and
palette, adding new menuitems.

(Includes a unit test)

[1] http://wiki.sugarlabs.org/go/Features/Web_services