bundleregistry: Add asynchronous install capabilities
In preparation for automatic activity updates, add asynchronous install
capabilities to the bundleregistry.

The synchronous install implementation has been removed and is now a user
of the async API that waits for completion before returning.

In order to avoid painful corner cases, the bundleregistry maintains
a queue of bundles to install, and installs them one by one. Install is
done in another thread to not block the main one.

As the install thread requires brief access to the list of bundles in the
bundle registry, all accesses to bundleregistry._bundles are now serialized
with a lock.