Use extensions.py to manage extensions paths and imports
The intervention here is to set the stage for supporting installing
Sugar extensions in ~/.sugar. A new module, jarabe/extensions.py has
been added which manages the location of Sugar extensions and the
import of the modules used by the extensions. (Previously,
config.ext_path was used to define the location of the extensions, but
now the location(s) are constructed from config.ext_path and
(eventually) ~/.sugar.) The various uses of __import__ throughout the
extension code has been replaced by extensions.load_module (which uses
imp.find_module and imp.load_module).

One somewhat unrelated change is to rename jarabe/web to
jarabe/webservice so as to help disambiguate the webservice modules
from the HTML5/Javascript work.

Note that there are some namespace collisions that impact that we are
still sorting through, but this code seems quite solid.