Add menuitems for webservices to jounal palettes
This patch adds UI support for interacting with online accounts from Sugar.
The relevant feature request is [1].

The new UI elements are added to the Sugar Journal Detail View toolbar:
(1) webservices are added to the Journal list view Copy-to palette and
to the similar palette on the Journal detail view toolbar.
(2) A palette of refresh buttons for all webservices are also added to
the Journal detail view toolbar. These buttons are enabled when the
webservice is active and insensitive when the webservice is inactive.
It also adds a new UI element to the Sugar List View palette: webservices
are added to the Journal list view Copy-to palette.

[1] http://wiki.sugarlabs.org/go/Features/Web_services