Don't initialise GTK+/dbus thread support
These calls allow GTK+ and dbus to be used from other threads.
Sugar doesn't currently create any threads, so these calls aren't
needed.

This is useful for followup work where some main loop recursion
results in a function that is sometimes called with the GDK threads lock
held, and sometimes without it held. By disabling this threading support,
this lock (and the surrounding headache) goes away.

By not activating the threading lock, we should also have a minor
performance increase.

Threads can still be used if desired, but all GTK+ calls should be made
from the main thread. These functions have actually been deprecated
in GTK+ 3.6 as the developers have determined that "draw from one thread"
is the way to go.

The GLib.threads_init() call is still needed, as we do have users
that work with threads, so the GIL must be released accordingly. However it
can be removed once pygobject-3.10 is a requirement:
https://bugzilla.gnome.org/show_bug.cgi?id=686914