Automatic activity updates
A new auto_update_frequency gconf key (default 0, disabled) enables
automatic activity update at an interval specified in days.

Sugar checks if this threshold has been met 10 minutes after launch,
and then every hour.

Activity upgrades are done completely in the background, but if the
user opens the updater control panel section while an update is in
progress, the background update is effectively brought into the
foreground and the user can monitor its progress.

This required a bit of view/model rework so that the view can be regarded
as a client that can attach at any point during the update process.