Implement urgent background activity updates
Dropping an empty .sugar-update flag file in the home directory will
now trigger an automatic background activity update ran as soon as
network connectivity is available.

This is designed for use after a system upgrade. Upon major Sugar and OS
upgrades, user activities will often need to be updated to new APIs, etc.
The system upgrade tool should put this file in place, which will then
trigger Sugar to try hard to update the user's activities.

This replaces the previous use of the .sugar-update functionality,
which was to show a message advising activity update, which turned out
to be somewhat ineffective in the field.